backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
# language-tool-python==2.7.1  # Requires Java
transformers==4.35.2
torch>=2.2.0
python-dotenv==1.0.0
requests==2.31.0